complete assignment 4

+++ NopBai/JAVA/assignment_4/Car.java
@@ -12,7 +12,7 @@ public class Car {
 		this.color = color;
 	}
 	
-	double getSalePrice(double salePrice){
+	int getSalePrice(int salePrice){
 		return(salePrice);
 	}
